add condition to calling sub deleteDbRef
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/GUS/trunk@1388 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ Common/plugin/perl/InsertDbRefAndDbRefNASequence.pm
@@ -41,6 +41,10 @@ sub new {
      {o => 'pattern2',
       t => 'string',
       h => 'identifier pattern for the na_sequence_id stored in the DbRefNASequence table, with parenthesis, e.g. \s+DT.(\d+)',
+     },
+     {o => 'delete',
+      t => 'boolean',
+      h => 'option to delete entries in dbref that are not in the current mapping',
      }
     ];
 
@@ -83,7 +87,7 @@ sub run {
 
   $self->insertDBRefNASeq($sourceIdDbrefHash,$mapHash); 
 
-  $self->deleteDbRef($dbRefHash);
+  $self->deleteDbRef($dbRefHash) if $self->getArgs()->{'delete'};
 }
 
 sub getSourceIdsAndMap {
